Completed the challenge and added UUID functionalities
The functions handleToggleTaskCompletion(), handleRemoveTask() and
handleToggleTaskCompletion() have been implemented. All three of
them do their functions by creating a list of tasks out of the page
state, modifying the list in the intended way and then setting that
list to be the new program state.

Furthermore, a UUID package was added to the application in order
to make each task id unique.